Created toml utils
